+++ app/src/main/java/intern/line/me/kyotoaclient/GetUserProfileActivity.kt
@@ -2,15 +2,9 @@ package intern.line.me.kyotoaclient
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.TextView
-import com.google.gson.FieldNamingPolicy
-import com.google.gson.GsonBuilder
 import intern.line.me.kyotoaclient.lib.model.User
-import intern.line.me.kyotoaclient.lib.api.GetUserInfo
-import intern.line.me.kyotoaclient.lib.api.interfaces.UserAPI
+import intern.line.me.kyotoaclient.lib.api.presenter.GetUserInfo
 import kotlinx.android.synthetic.main.activity_get_user_profile.*
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 
 class GetUserProfileActivity : AppCompatActivity() {
