+++ app/src/main/java/intern/line/me/kyotoaclient/lib/api/adapters/MessagesAdapter.kt
@@ -6,10 +6,10 @@ import android.widget.Toast
 import intern.line.me.kyotoaclient.MessageActivity
 import intern.line.me.kyotoaclient.lib.model.Message
 import intern.line.me.kyotoaclient.lib.model.MessageList
-import intern.line.me.kyotoaclient.lib.api.CreateMessage
-import intern.line.me.kyotoaclient.lib.api.DeleteMessage
-import intern.line.me.kyotoaclient.lib.api.GetMessages
-import intern.line.me.kyotoaclient.lib.api.UpdateMessage
+import intern.line.me.kyotoaclient.lib.api.presenter.CreateMessage
+import intern.line.me.kyotoaclient.lib.api.presenter.DeleteMessage
+import intern.line.me.kyotoaclient.lib.api.presenter.GetMessages
+import intern.line.me.kyotoaclient.lib.api.presenter.UpdateMessage
 import kotlinx.coroutines.experimental.Job
 import kotlinx.coroutines.experimental.launch
